% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_99_app.R
\name{cruncher_server}
\alias{cruncher_server}
\title{Server}
\usage{
cruncher_server(id, project_dir, annot_dir, ...)
}
\arguments{
\item{id}{id}

\item{project_dir}{project dir}

\item{annot_dir}{the directory for annotation databases (global), the project specific annotation is in the .gsannot folder}

\item{...}{not used currently}

\item{userinfo}{the user information, including - dir, type, typeEnd, addno}

\item{demoData}{path to the demo data file}
}
\description{
Server
}
