% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_02_normalization.R
\name{module_normalization}
\alias{module_normalization}
\title{importFrom shinycssloaders withSpinner}
\usage{
module_normalization(id, obj)
}
\description{
importFrom shinycssloaders withSpinner
}
