% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_prepESVObj.R
\name{prepESVObj}
\alias{prepESVObj}
\title{Internal function to create object for ESV}
\usage{
prepESVObj(x)
}
\arguments{
\item{x}{configured object}
}
\description{
Internal function to create object for ESV
}
